Sync Package: impemented OrderableSyncPackageWriter to fix bug in generating sync package sort index. KP-4909

+++ src/Core/Synchronization/WB.Core.Synchronization/Implementation/ReadSide/Indexes/InterviewSyncPackagesByBriefFields.cs
@@ -19,7 +19,7 @@ namespace WB.Core.Synchronization.Implementation.ReadSide.Indexes
 
             public Guid UserId { get; set; }
 
-            public int SortIndex { get; set; }
+            public long SortIndex { get; set; }
         }
 
         public InterviewSyncPackagesByBriefFields()
@@ -34,7 +34,7 @@ namespace WB.Core.Synchronization.Implementation.ReadSide.Indexes
                        };
 
             this.Index(x => x.SortIndex, FieldIndexing.Default);
-            this.Sort(x => x.SortIndex, SortOptions.Int);
+            this.Sort(x => x.SortIndex, SortOptions.Long);
         }
     }
 }
