+++ src/Core/Synchronization/WB.Core.Synchronization/SyncStorage/IIndexedView.cs
@@ -4,6 +4,6 @@ namespace WB.Core.Synchronization.SyncStorage
 {
     public interface IIndexedView : IView
     {
-        int SortIndex { get; }
+        long SortIndex { get; }
     }
 }
